requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
jinja2==3.1.2
python-dotenv==1.0.0
python-nmap==0.7.1
